Update SDL2 to SDL3
Update SDL2 to SDL3 and all child libraries of sdl

Code adapted to SDL3

Add libuv (for network)

+++ Source/Public/Assets/Assets/TextureAsset.h
@@ -21,7 +21,7 @@ public:
 	/** End FAssetBase */
 
 	/** Returns size of texture. */
-	FVector2D<int> GetSize() const;
+	FVector2D<float> GetSize() const;
 
 	/** Get texture */
 	FTexture* GetTexture() const;
